+++ CHANGELOG.md
@@ -4,6 +4,7 @@ New safe operations
 
 - Changing between `text` and `citext` when not indexed
 - Changing a `string` column to a `citext` column when not indexed
+- Changing a `citext` column to a `string` column with no length limit
 
 ## 0.3.0 (2022-02-10)